Validate host address on startup
