+++ comm/src/main/scala/coop/rchain/comm/package.scala
@@ -1,7 +1,12 @@
 package coop.rchain
 
+import java.net.InetAddress
+
+import scala.util.Try
+
 import cats.mtl.ApplicativeAsk
 
+import coop.rchain.catscontrib.ski.kp
 import coop.rchain.metrics.Metrics
 
 package object comm {
@@ -12,4 +17,20 @@ package object comm {
   }
 
   val CommMetricsSource: Metrics.Source = Metrics.Source(Metrics.BaseSource, "comm")
+
+  def isValidInetAddress(host: String): Boolean =
+    Try(InetAddress.getByName(host))
+      .fold(kp(false), !_.isAnyLocalAddress)
+
+  def isValidPublicInetAddress(host: String): Boolean =
+    Try(InetAddress.getByName(host))
+      .fold(
+        kp(false),
+        a =>
+          !(a.isAnyLocalAddress ||
+            a.isLinkLocalAddress ||
+            a.isLoopbackAddress ||
+            a.isMulticastAddress ||
+            a.isSiteLocalAddress)
+      )
 }
