+++ integration-tests/test/rnode.py
@@ -511,6 +511,8 @@ def make_bootstrap_node(
     container_command_flags = set([
         "--standalone",
         "--prometheus",
+        "--no-upnp",
+        "--allow-private-addresses"
     ])
 
     container_command_options = {
@@ -587,6 +589,8 @@ def make_peer(
 
     container_command_flags = set([
         "--prometheus",
+        "--no-upnp",
+        "--allow-private-addresses"
     ])
 
     if cli_flags is not None:
